takt: fix-empty-change-reason
